Split out BorrowingFieldsVisitor
